+++ internal/ecosystem/cratesio/const.go
@@ -1,0 +1,7 @@
+package cratesio
+
+const (
+	EcosystemName = "crates.io"
+	CargoFile     = "Cargo.toml"
+	CargoLockFile = "Cargo.lock"
+)