+++ internal/ecosystem/packagist/detection.go
@@ -1,20 +1,14 @@
 package packagist
 
 import (
-	"github.com/1franck/cvepack/internal/common"
-	"path/filepath"
+    "github.com/1franck/cvepack/internal/common"
+    "path/filepath"
 )
 
 func DetectComposerJson(path string) bool {
-	if common.FileExists(filepath.Join(path, ComposerFile)) {
-		return true
-	}
-	return false
+    return common.FileExists(filepath.Join(path, ComposerFile))
 }
 
 func DetectComposerLock(path string) bool {
-	if common.FileExists(filepath.Join(path, ComposerLockFile)) {
-		return true
-	}
-	return false
+    return common.FileExists(filepath.Join(path, ComposerLockFile))
 }